fix(inventory) : 인벤토리 디렉토리 수정 #19
